The thomas branch

+++ server/functions/handlers/posts.js
@@ -102,6 +102,9 @@ exports.getPost = (req,res) => {
 
 exports.deletePost = (req,res) => {
   db.doc(`/posts/${req.params.postId}`).get().then(doc => {
+    if(!doc.exists){
+      return res.status(404).json({error: "Could not find post."})
+    }
     let post = doc.data()
     if(post.authorid === req.user.id){
       doc.delete().then(() => {
@@ -111,7 +114,41 @@ exports.deletePost = (req,res) => {
           return res.json({success: "Successfully deleted post"})
         }
         // delete comments and likes and stuff
+      }).catch(err => {
+        console.error(err)
+        return res.status(500).json({error: "Error deleting post"})
       })
+    } else {
+      console.log(post.authorid, req.user.id)
+      return res.status(403).json({error: "Post isnt yours lol"})
     }
+  }).catch(err => {
+    console.error(err)
+    return res.status(500).json({error: "Could not get post"})
+  })
+}
+
+exports.editPost = (req,res) => {
+  db.doc(`/posts/${req.params.postId}`).get().then(doc => {
+    if(!doc.exists){
+      return res.status(404).json({error: "Post does not exist"})
+    }
+    if(doc.data().authorid === req.user.id){
+      doc.ref.update(req.body.update).then(() => {
+        if(req.auth.refreshed){
+          return res.json({success: "Successfully edited post", refreshed: true, token: req.auth.token, expires: req.auth.expires})
+        } else {
+          return res.json({success: "Successfully edited post"})
+        }
+      }).catch(err => {
+        console.error(err)
+        return res.status(500).json({error: "Error updating post."})
+      })
+    } else {
+      return res.status(403).json({error: "post isnt yours lol"})
+    }
+  }).catch(err => {
+    console.error(err)
+    return res.status(500).json({error: "Error getting post"})
   })
 }
